requirements.txt
fastapi
uvicorn
pydantic
streamlit
pandas
numpy
python-multipart
requests
pytest
httpx
casadi
scipy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-client
python-jose[cryptography]
passlib[bcrypt]
asyncua
fpdf2
matplotlib
reportlab
PyYAML
filelock==3.12.3
sqlalchemy
alembic
psycopg2-binary
sqlite-utils
PyJWT[crypto]
